requirements.txt
flask
flask-socketio
pymongo
python-dotenv
pyjwt
